third_party/WebKit/LayoutTests/accessibility/inline-text-change-style-expected.txt

This paragraph contains two lines of text.

Tests that accessible inline text boxes update when the static text style changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axStaticText.childrenCount is 2
PASS axInlineBefore0.name is 'This paragraph contains '
PASS axInlineBefore1.name is 'two lines of text.'
PASS axStaticText.childrenCount is 1
PASS axInlineAfter0.name is 'This paragraph contains two lines of text.'
PASS successfullyParsed is true

TEST COMPLETE
